network-error-logging/sends-report-on-success.https.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test that NEL reports are sent for successful requests</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='./support/fetch.sub.js'></script>
</head>
<body>
  <script>
    promise_test(async t => {
      await fetchNELPolicyResource();
      assert_true(await reportExists({
        url: getNELPolicyResourceURL(),
        type: "network-error",
        report: {
          uri: getNELPolicyResourceURL(),
          sampling_fraction: 1.0,
          status_code: 200,
          type: "ok",
        },
      }));
    });
  </script>
</body>
</html>
